ADD: solución primer problema
