+++ src/index.html
@@ -1,0 +1,14 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>Reto 3</title>
+</head>
+<body>
+    <script src="primerProblema.js"></script>
+    <!-- <script src="segundoProblema.js"></script>
+    <script src="tercerProblema.js"></script> -->
+</body>
+</html>
